Add core functionality for analytics, Kinsta integration, and user management
- Implemented analytics routes to fetch and display domain statistics.
- Created Kinsta API routes for cache clearing, plugin, and theme retrieval.
- Developed core routes for domain management and credential storage.
- Added dashboard functionality to summarize domain data.
- Introduced scraping routes to manage scraping tasks and trigger Scrapy spiders.
- Enhanced user management with bulk user addition and API endpoints for user operations.
- Established database utility functions for MySQL connections and queries.
- Integrated visualization features using Plotly for data representation.
- Configured MySQL connection settings in a separate config file.
